+++ examples/demo/src/controllers/auth.rs
@@ -1,4 +1,4 @@
-use loco_rs::prelude::*;
+use loco_rs::{controller::bad_request, prelude::*};
 use serde::{Deserialize, Serialize};
 
 use crate::{
@@ -7,7 +7,7 @@ use crate::{
         _entities::users,
         users::{LoginParams, RegisterParams},
     },
-    views::auth::LoginResponse,
+    views::auth::UserSession,
 };
 #[derive(Debug, Deserialize, Serialize)]
 pub struct VerifyParams {
@@ -30,18 +30,16 @@ pub struct ResetParams {
 async fn register(
     State(ctx): State<AppContext>,
     Json(params): Json<RegisterParams>,
-) -> Result<Json<()>> {
+) -> impl IntoResponse {
     let res = users::Model::create_with_password(&ctx.db, &params).await;
 
     let user = match res {
         Ok(user) => user,
         Err(err) => {
-            tracing::info!(
-                message = err.to_string(),
-                user_email = &params.email,
-                "could not register user",
-            );
-            return format::json(());
+            let msg = "could not register user";
+
+            tracing::info!(message = err.to_string(), user_email = &params.email, msg,);
+            return bad_request(msg);
         }
     };
 
@@ -52,7 +50,12 @@ async fn register(
 
     AuthMailer::send_welcome(&ctx, &user).await?;
 
-    format::json(())
+    let jwt_secret = ctx.config.get_jwt_config()?;
+
+    let token = user
+        .generate_jwt(&jwt_secret.secret, &jwt_secret.expiration)
+        .or_else(|_| unauthorized("unauthorized!"))?;
+    format::json(UserSession::new(&user, &token))
 }
 
 /// Verify register user. if the user not verified his email, he can't login to
@@ -60,7 +63,7 @@ async fn register(
 async fn verify(
     State(ctx): State<AppContext>,
     Json(params): Json<VerifyParams>,
-) -> Result<Json<()>> {
+) -> impl IntoResponse {
     let user = users::Model::find_by_verification_token(&ctx.db, &params.token).await?;
 
     if user.email_verified_at.is_some() {
@@ -71,7 +74,7 @@ async fn verify(
         tracing::info!(pid = user.pid.to_string(), "user verified");
     }
 
-    format::json(())
+    format::empty_json()
 }
 
 /// In case the user forgot his password  this endpoints generate a forgot token
@@ -81,11 +84,11 @@ async fn verify(
 async fn forgot(
     State(ctx): State<AppContext>,
     Json(params): Json<ForgotParams>,
-) -> Result<Json<()>> {
+) -> impl IntoResponse {
     let Ok(user) = users::Model::find_by_email(&ctx.db, &params.email).await else {
         // we don't want to expose our users email. if the email is invalid we still
         // returning success to the caller
-        return format::json(());
+        return format::empty_json();
     };
 
     let user = user
@@ -95,30 +98,33 @@ async fn forgot(
 
     AuthMailer::forgot_password(&ctx, &user).await?;
 
-    format::json(())
+    format::empty_json()
 }
 
 /// reset user password by the given parameters
-async fn reset(State(ctx): State<AppContext>, Json(params): Json<ResetParams>) -> Result<Json<()>> {
+async fn reset(
+    State(ctx): State<AppContext>,
+    Json(params): Json<ResetParams>,
+) -> impl IntoResponse {
     let Ok(user) = users::Model::find_by_reset_token(&ctx.db, &params.token).await else {
         // we don't want to expose our users email. if the email is invalid we still
         // returning success to the caller
         tracing::info!("reset token not found");
 
-        return format::json(());
+        return format::empty_json();
     };
     user.into_active_model()
         .reset_password(&ctx.db, &params.password)
         .await?;
 
-    format::json(())
+    format::empty_json()
 }
 
 /// Creates a user login and returns a token
 async fn login(
     State(ctx): State<AppContext>,
     Json(params): Json<LoginParams>,
-) -> Result<Json<LoginResponse>> {
+) -> impl IntoResponse {
     let user = users::Model::find_by_email(&ctx.db, &params.email).await?;
 
     let valid = user.verify_password(&params.password);
@@ -133,7 +139,7 @@ async fn login(
         .generate_jwt(&jwt_secret.secret, &jwt_secret.expiration)
         .or_else(|_| unauthorized("unauthorized!"))?;
 
-    format::json(LoginResponse::new(&user, &token))
+    format::json(UserSession::new(&user, &token))
 }
 
 pub fn routes() -> Routes {