+++ examples/demo/src/mailers/auth/welcome/html.t
@@ -4,7 +4,7 @@
   Dear {{name}},
   Welcome to Loco! You can now log in to your account.
   Before you get started, please verify your account by clicking the link below:
-  <a href="http://{{domain}}/verify#{{verifyToken}}">
+  <a href="{{host}}/auth/verify/{{verifyToken}}">
     Verify Your Account
   </a>
   <p>Best regards,<br>The Loco Team</p>