+++ examples/demo/src/models/users.rs
@@ -1,10 +1,9 @@
 use async_trait::async_trait;
 use chrono::offset::Local;
 use loco_rs::{
-    auth,
-    concern::query::prelude::*,
-    hash,
+    auth, hash,
     model::{Authenticable, ModelError, ModelResult},
+    prelude::model::query,
     validation,
     validator::Validate,
 };
@@ -81,14 +80,18 @@ impl ActiveModelBehavior for super::_entities::users::ActiveModel {
 impl Authenticable for super::_entities::users::Model {
     async fn find_by_api_key(db: &DatabaseConnection, api_key: &str) -> ModelResult<Self> {
         let user = users::Entity::find()
-            .filter(condition().eq(users::Column::ApiKey, api_key).build())
+            .filter(
+                query::dsl::condition()
+                    .eq(users::Column::ApiKey, api_key)
+                    .build(),
+            )
             .one(db)
             .await?;
         user.ok_or_else(|| ModelError::EntityNotFound)
     }
 
     async fn find_by_claims_key(db: &DatabaseConnection, claims_key: &str) -> ModelResult<Self> {
-        super::_entities::users::Model::find_by_pid(db, claims_key).await
+        Self::find_by_pid(db, claims_key).await
     }
 }
 
@@ -100,7 +103,11 @@ impl super::_entities::users::Model {
     /// When could not find user by the given token or DB query error
     pub async fn find_by_email(db: &DatabaseConnection, email: &str) -> ModelResult<Self> {
         let user = users::Entity::find()
-            .filter(condition().eq(users::Column::Email, email).build())
+            .filter(
+                query::dsl::condition()
+                    .eq(users::Column::Email, email)
+                    .build(),
+            )
             .one(db)
             .await?;
         user.ok_or_else(|| ModelError::EntityNotFound)
@@ -117,7 +124,7 @@ impl super::_entities::users::Model {
     ) -> ModelResult<Self> {
         let user = users::Entity::find()
             .filter(
-                condition()
+                query::dsl::condition()
                     .eq(users::Column::EmailVerificationToken, token)
                     .build(),
             )
@@ -133,7 +140,11 @@ impl super::_entities::users::Model {
     /// When could not find user by the given token or DB query error
     pub async fn find_by_reset_token(db: &DatabaseConnection, token: &str) -> ModelResult<Self> {
         let user = users::Entity::find()
-            .filter(condition().eq(users::Column::ResetToken, token).build())
+            .filter(
+                query::dsl::condition()
+                    .eq(users::Column::ResetToken, token)
+                    .build(),
+            )
             .one(db)
             .await?;
         user.ok_or_else(|| ModelError::EntityNotFound)
@@ -171,7 +182,11 @@ impl super::_entities::users::Model {
     pub async fn find_by_pid(db: &DatabaseConnection, pid: &str) -> ModelResult<Self> {
         let parse_uuid = Uuid::parse_str(pid).map_err(|e| ModelError::Any(e.into()))?;
         let user = users::Entity::find()
-            .filter(condition().eq(users::Column::Pid, parse_uuid).build())
+            .filter(
+                query::dsl::condition()
+                    .eq(users::Column::Pid, parse_uuid)
+                    .build(),
+            )
             .one(db)
             .await?;
         user.ok_or_else(|| ModelError::EntityNotFound)
@@ -184,7 +199,11 @@ impl super::_entities::users::Model {
     /// When could not find user by the given token or DB query error
     pub async fn find_by_api_key(db: &DatabaseConnection, api_key: &str) -> ModelResult<Self> {
         let user = users::Entity::find()
-            .filter(condition().eq(users::Column::ApiKey, api_key).build())
+            .filter(
+                query::dsl::condition()
+                    .eq(users::Column::ApiKey, api_key)
+                    .build(),
+            )
             .one(db)
             .await?;
         user.ok_or_else(|| ModelError::EntityNotFound)
@@ -209,7 +228,11 @@ impl super::_entities::users::Model {
         let txn = db.begin().await?;
 
         if users::Entity::find()
-            .filter(condition().eq(users::Column::Email, &params.email).build())
+            .filter(
+                query::dsl::condition()
+                    .eq(users::Column::Email, &params.email)
+                    .build(),
+            )
             .one(&txn)
             .await?
             .is_some()