+++ examples/demo/src/views/dashboard.rs
@@ -1,6 +1,11 @@
 use loco_rs::prelude::*;
 use serde_json::json;
 
-pub fn home(v: impl ViewRenderer) -> Result<impl IntoResponse> {
-    format::render().view(&v, "home/hello.html", json!({}))
+/// Home view
+///
+/// # Errors
+///
+/// This function will return an error if render fails
+pub fn home(v: &impl ViewRenderer) -> Result<impl IntoResponse> {
+    format::render().view(v, "home/hello.html", json!({}))
 }