+++ src/gen/mod.rs
@@ -41,6 +41,13 @@ const DEPLOYMENT_OPTIONS: &[(&str, DeploymentKind)] = &[
     ("Nginx", DeploymentKind::Nginx),
 ];
 
+#[derive(clap::ValueEnum, Clone, Debug)]
+pub enum ScaffoldKind {
+    Api,
+    Html,
+    Htmx,
+}
+
 #[derive(Debug, Clone)]
 pub enum DeploymentKind {
     Docker,
@@ -86,6 +93,9 @@ pub enum Component {
 
         /// Model and params fields, eg. title:string hits:int
         fields: Vec<(String, String)>,
+
+        // k
+        kind: ScaffoldKind,
     },
     Controller {
         /// Name of the thing to generate
@@ -105,10 +115,9 @@ pub enum Component {
     },
     Deployment {},
 }
-
+#[allow(clippy::too_many_lines)]
 pub fn generate<H: Hooks>(component: Component, config: &Config) -> Result<()> {
     let rrgen = RRgen::default();
-
     match component {
         #[cfg(feature = "with-db")]
         Component::Model {
@@ -123,8 +132,11 @@ pub fn generate<H: Hooks>(component: Component, config: &Config) -> Result<()> {
             );
         }
         #[cfg(feature = "with-db")]
-        Component::Scaffold { name, fields } => {
-            println!("{}", scaffold::generate::<H>(&rrgen, &name, &fields)?);
+        Component::Scaffold { name, fields, kind } => {
+            println!(
+                "{}",
+                scaffold::generate::<H>(&rrgen, &name, &fields, &kind)?
+            );
         }
         #[cfg(feature = "with-db")]
         Component::Migration { name } => {